requirements.txt
flask
flask_restful
flask-sqlalchemy
pymysql
flask-migrate
flask-script
flask-jwt-extended
requests
flask-cors
password_strength
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client

